tests/requirements.txt
aiohttp==3.8.1
async-timeout==4.0.3
attrs>=17.4.0
chardet==3.0.4
elastic-transport~=8.17.0
elasticsearch[async]==8.13.2
email-validator>=2.0.0
faker==19.2.0
fastapi-cli>=0.0.2
fastapi==0.111.0
httpx>=0.23.0
importlib-metadata==6.8.0
iniconfig
jinja2>=2.11.2
more-itertools~=10.6.0
multidict==4.5
orjson==3.8.12
packaging
pluggy==1.6.0
py>=1.8.2
pydantic-settings==2.8.1
pydantic~=2.10.6
pytest-asyncio==1.0.0
pytest==8.4.1
python-dateutil>=2.4
python-decouple==3.8
python-dotenv>=0.21.0
python-multipart>=0.0.7
redis==4.5.3
requests==2.32.4
starlette==0.37.2
tenacity==9.0.0
toml==0.10.2
typing-extensions>=4.8.0
ujson!=4.0.2,!=4.1.0,!=4.2.0,!=4.3.0,!=5.0.0,!=5.1.0,>=4.0.1
uvicorn[standard]>=0.12.0
SQLAlchemy==2.0.39
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0